+++ piecrust/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `Session::memory_len` to get the length of a memory in session [#268]
+
 ### Changed
 
 - Change minimum number of pages to be 4
@@ -16,6 +20,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Fake guard pages are now removed
+
+### Fixed
+
+- Revert memory size on errors [#268]
+- Fix reporting of memory size to `wasmer` [#268]
 
 ## [0.9.3] - 2023-09-07
 
@@ -218,6 +227,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#234]: https://github.com/dusk-network/piecrust/pull/234
 
 <!-- ISSUES -->
+[#268]: https://github.com/dusk-network/piecrust/issues/268
 [#254]: https://github.com/dusk-network/piecrust/issues/254
 [#253]: https://github.com/dusk-network/piecrust/issues/253
 [#249]: https://github.com/dusk-network/piecrust/issues/249